requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
django-decouple==2.1
django-filter==24.2
django-import-export==4.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-tinymce-4
pandas
openpyxl
python-slugify==8.0.4
